requirements.txt
APScheduler==3.10.0
asgiref==3.6.0
atatus==1.4.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
configparser==5.3.0
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-debug-toolbar==3.2.1
django-grappelli==3.0.4
django-guardian==2.4.0
django-mmc==1.6.5
djangorestframework==3.14.0
Faker==17.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.2
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
translate==3.6.1
treads==1.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
